+++ pkg/codegen/importer/hcl2.go
@@ -425,9 +425,17 @@ func generateValue(typ schema.Type, value resource.PropertyValue) (model.Express
 				if err != nil {
 					return nil, err
 				}
+
+				// we need to take into account when we have a complex key - without this
+				// we will return key as an array as the / will show as 2 parts
+				propKey := string(k)
+				if strings.Contains(string(k), "/") {
+					propKey = fmt.Sprintf("%q", string(k))
+				}
+
 				items = append(items, model.ObjectConsItem{
 					Key: &model.LiteralValueExpression{
-						Value: cty.StringVal(string(k)),
+						Value: cty.StringVal(propKey),
 					},
 					Value: x,
 				})